2010-05-26  Noel O'Boyle
	* Lots of files: 'Fix' warnings with 64-bit Windows relating to size_t
	and unsigned ints. I either:
	(a) changed unsigned int to size_t where the function was GetSize() or
	Size() or something
	or...in every other case...
	(b) static_cast<unsigned int> the size_t value (for example, no-one
	expects NumResidues() to be size_t, or GetValence()

+++ include/openbabel/atom.h
@@ -231,7 +231,7 @@ namespace OpenBabel
       //! \return The current number of explicit connections
       unsigned int GetValence()       const
         {
-          return((_vbond.empty()) ? 0 : _vbond.size());
+          return((_vbond.empty()) ? 0 : static_cast<unsigned int> (_vbond.size()));
         }
       //! \return The hybridization of this atom: 1 for sp, 2 for sp2, 3 for sp3, 4 for sq. planar, 5 for trig. bipy, 6 for octahedral
       unsigned int GetHyb()             const;
