+++ include/openbabel/bitvec.h
@@ -118,7 +118,7 @@ namespace OpenBabel
       /// Return the bit offset of the last bit (for iterating) i.e. -1
       int EndBit() const {  return -1; }
       /// Return the number of words ( NOT the number of bits ).
-      unsigned GetSize() const    { return(_size);    }
+      size_t GetSize() const    { return(_size);    }
       /// Return the number of bits which are set to 1 in the vector
       unsigned CountBits() const;
