+++ include/openbabel/bondtyper.h
@@ -39,7 +39,7 @@ public:
     //@{
     void ParseLine(const char*);
     //! \return the size of the database (for error checking)
-    unsigned int GetSize()                 { return _fgbonds.size();}
+    size_t GetSize()                 { return _fgbonds.size();}
     //@}
     
     //! \name Bond Perception Routines
